refactoring(Icons): replace remaining icons with new Icon component
