+++ src/components/typography/typography.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { up } from '../style-utils/breakpoint';
 import { theme } from '../layout/theme';
-import { Swoosh } from '../icons/swoosh';
+import { Swoosh } from '../header/icons/swoosh';
 import { HEADER_HEIGHT } from '../header/header';
 import { Link } from '../links/links';
